requirements.txt
openai
streamlit
azure-cognitiveservices-speech